added image storage inside database
stores images on outside database called cloudinary. The link to the
image is stored in our database. There needs to be css cleanup on the
forms relating to edit event and create event, and there needs to be css
written to display the image for the event on the event lists.